Change data types to get a smaller memory footprint
git-svn-id: svn://svn.cc65.org/cc65/trunk@2595 b7a2c559-68d2-44c3-8de9-860c34a00d81

+++ src/common/filepos.h
@@ -46,10 +46,10 @@
 
 /* Type of a file position */
 typedef struct FilePos FilePos;
-struct FilePos {     
+struct FilePos {
     unsigned long   Line;    	       	/* Line */
-    unsigned 	    Col;		/* Column */
-    unsigned 	    Name;		/* File */
+    unsigned short  Col;		/* Column */
+    unsigned short  Name;		/* File */
 };
